v1.8.4 which is fully Py3/Py2 build compatible.
This works with Py2/5.8.0-109 and builds in Py3/CASA-6.4.0-16-py3.8.

The 'buildmytasks' paradigm is still broken in Py3 in that the
descriptions of parameters do not survive into the inp arena, but
at some point maybe they will fix that; on the other hand, no one
is now using that mode so it can go away eventually.



git-svn-id: https://svn.atnf.csiro.au/difx@10393 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876